pushing working data-getter function
